+++ tests/testthat/test-types.R
@@ -45,10 +45,10 @@ test_that("lang_type_of() returns correct lang subtype", {
   expect_identical(lang_type_of(quote(foo@bar())), "recursive")
 
   lang <- quote(foo())
-  mut_node_car(lang, 10)
+  node_poke_car(lang, 10)
   expect_error(lang_type_of(lang), "corrupt")
 
-  mut_node_car(lang, base::list)
+  node_poke_car(lang, base::list)
   expect_identical(lang_type_of(lang), "inlined")
 })